Updating dependencies for Ubuntu 21 (#263)
- New clang dependency (10.0.1)
- New boost depdendecy (1.65.1)
- Removed 'clang-10' from 'scripts/install_apt_deps.sh'
- Removed 'libboost-all-dev' from 'scripts/install_apt_deps.sh'
- Fixed deps configuraton detection
- Fixed in conda environment creating in 'Dockerfile'

+++ install_deps.sh
@@ -27,6 +27,7 @@ help_msg="ILLIXR Dependency Installer:
 
   ## Yes to all dependencies and default value prompts
     -y/--yes
+  ** WARNING ** This option may cause this script to _DELETE_ data without supervision
 
   ## Number of jobs/cores/threads for make to use (default: '${illixr_nproc}')
     -j/--jobs <integer>
@@ -103,9 +104,15 @@ function prompt_install()
 
         export src_dir="${parent_dir}/${dep_name}"
 
-        detect_dependency "${dep_name}" "${deps_log_dir}" "${enable_quiet}"
+        local enable_dry_run="yes"
+        detect_dependency "${dep_name}" "${deps_log_dir}" "${enable_dry_run}"
         if [ "$?" -eq 0 ]; then
             echo "Detected previous installation for '${dep_name}'."
+
+            if [ "${enable_quiet}" = "no" ]; then
+                tail --lines=5 "${dep_log_path}"
+            fi
+
             if y_or_n "Use the configuration in '${dep_log_path}'?"; then
                 ## Load the configuration
                 . "${dep_log_path}"
@@ -119,11 +126,16 @@ function prompt_install()
 
         if [ -d "${src_dir}" ]; then
             echo "Source directory '${src_dir}' already exists."
+
             if y_or_n "Clear directory and proceed with installation?"; then
                 rm -rf --preserve-root "${src_dir}"
+
                 if [ "$?" -eq 1 ]; then
-                    echo "Failed to clear '${src_dir}. Trying with sudo (if interactive, Ctrl+c to abort)."
-                    sudo rm -rf --preserve-root "${src_dir}"
+                    echo "Failed to clear '${src_dir}."
+
+                    if y_or_n "Try to clear '${src_dir}' with sudo?"; then
+                        sudo rm -rf --preserve-root "${src_dir}"
+                    fi
                 fi
             else
                 return 1
@@ -207,175 +219,200 @@ fi
 
 ### Main ###
 
-if [ "${ID_LIKE}" = debian ] || [ "${ID}" = debian ]; then
-    ## For system-wide or from-source installs that are not possible via apt
-    sudo mkdir -p "${opt_dir}"
-    sudo chown "${USER}:" "${opt_dir}"
-
-    ## Source the configurations for our dependencies
-    . deps.sh
-
-    echo "The user will now be prompted to install the following dependencies and optional features:"
-    echo "  Binary packages (via apt-get), Docker, CUDA, OpenCV, eigen, Vulkan, gtest, qemu, OpenXR-SDK, "
-    echo "  gtsam, opengv, DBoW2, Kimera-RPGO, Conda (miniconda3)"
-
-    if y_or_n "Add apt-get sources list/keys and install necessary packages"; then
-        if y_or_n "^^^^  Also install Docker (docker-ce) for local CI/CD debugging support"; then
-            export use_docker="yes"
-        fi
-        pmt_msg_warn_cuda="Also automate install of CUDA 11 (cuda) for GPU plugin support on Ubuntu (_only_!)"
-        pmt_msg_warn_cuda+="\n(This script will _not_ install the package on non-Ubuntu distributions, "
-        pmt_msg_warn_cuda+="or if a supported GPU is not found)"
-        if y_or_n "^^^^  ${pmt_msg_warn_cuda}"; then
-            export use_cuda="yes"
-        fi
-
-        if [ "${enable_quiet}" = "no" ]; then
-            . "${script_path_apt}"
-        else
-            . "${script_path_apt}" >/dev/null
-        fi
-    fi
-
-    prompt_install \
-        "${dep_name_opencv}" \
-        "${deps_log_dir}" \
-        "${script_path_opencv}" \
-        "${parent_dir_opencv}" \
-        "${dep_prompt_opencv}" \
-        "${dep_ver_opencv}"
-
-    prompt_install \
-        "${dep_name_vulkan}" \
-        "${deps_log_dir}" \
-        "${script_path_vulkan}" \
-        "${parent_dir_vulkan}" \
-        "${dep_prompt_vulkan}" \
-        "${dep_ver_vulkan}"
-
-    prompt_install \
-        "${dep_name_gtest}" \
-        "${deps_log_dir}" \
-        "${script_path_gtest}" \
-        "${parent_dir_gtest}" \
-        "${dep_prompt_gtest}" \
-        "${dep_ver_gtest}"
-
-    prompt_install \
-        "${dep_name_qemu}" \
-        "${deps_log_dir}" \
-        "${script_path_qemu}" \
-        "${parent_dir_qemu}" \
-        "${dep_prompt_qemu}" \
-        "${dep_ver_openxr}"
-
-    # if [ ! -d Vulkan-Loader ]; then
-    #   echo "Next: Install Vulkan Loader from source"
-    #   if y_or_n; then
-    #       git clone https://github.com/KhronosGroup/Vulkan-Loader.git
-    #       mkdir -p Vulkan-Headers/build && cd Vulkan-Headers/build
-    #       ../scripts/update_deps.py
-    #       cmake -C helper.cmake ..
-    #       cmake --build .
-    #       cd ../..
-    #   fi
-    # fi
-
-    prompt_install \
-        "${dep_name_openxr}" \
-        "${deps_log_dir}" \
-        "${script_path_openxr}" \
-        "${parent_dir_openxr}" \
-        "${dep_prompt_openxr}" \
-        "${dep_ver_openxr}"
-
-    prompt_install \
-        "${dep_name_gtsam}" \
-        "${deps_log_dir}" \
-        "${script_path_gtsam}" \
-        "${parent_dir_gtsam}" \
-        "${dep_prompt_gtsam}" \
-        "${dep_ver_gtsam}"
-
-    prompt_install \
-        "${dep_name_opengv}" \
-        "${deps_log_dir}" \
-        "${script_path_opengv}" \
-        "${parent_dir_opengv}" \
-        "${dep_prompt_opengv}" \
-        "${dep_ver_opengv}"
-
-    prompt_install \
-        "${dep_name_dbow2}" \
-        "${deps_log_dir}" \
-        "${script_path_dbow2}" \
-        "${parent_dir_dbow2}" \
-        "${dep_prompt_dbow2}" \
-        "${dep_ver_dbow2}"
-
-    prompt_install \
-        "${dep_name_kimera_rpgo}" \
-        "${deps_log_dir}" \
-        "${script_path_kimera_rpgo}" \
-        "${parent_dir_kimera_rpgo}" \
-        "${dep_prompt_kimera_rpgo}" \
-        "${dep_ver_kimera_rpgo}"
-
-    prompt_install \
-        "${dep_name_conda}" \
-        "${deps_log_dir}" \
-        "${script_path_conda}" \
-        "${parent_dir_conda}" \
-        "${dep_prompt_conda}" \
-        "${dep_ver_conda}"
-
-    ## Load new library paths
-    sudo ldconfig
-
-    ### Virtual environment creation ###
-
-    echo "Attempting to create a virtual environment configuration via conda ..."
-
-    ## Check for a previous conda installation
-    detect_dependency "${dep_name_conda}" "${deps_log_dir}"
-    if [ "$?" -eq 0 ]; then
-        dep_log_path_conda="${dep_log_path}"
-        src_dir_conda="${src_dir}"
-        echo "Found conda installation log '${dep_log_path_conda}' : dir <- '${src_dir_conda}'"
-    else
-        dep_log_path_conda="${dep_log_path}"
-        src_dir_conda="${parent_dir_conda}/${dep_name_conda}"
-
-        dep_missing_msg_conda="Installation log for conda not found at '${dep_log_path_conda}'."
-        dep_missing_msg_conda+="\n  Conda may have been installed without this script (or an older version)."
-
-        print_warning "${dep_missing_msg_conda}"
-
-        if ! y_or_n "Try to create a Python environment configuration anyway?"; then
-            echo "This was the last step. Exiting early."
-            exit 0
-        fi
-
-        ## Conda may have been installed without this script
-        echo "Assuming : dir <- '${src_dir_conda}'"
-    fi
-
-    env_config_parent_dir = $(dirname "${env_config_path}")
-    if [ ! -d "${env_config_parent_dir}" ]; then
-        mkdir -p "${env_config_parent_dir}"
-    fi
-
-    cmd_conda="${src_dir_conda}/bin/conda"
-    if [ -f "${cmd_conda}" ]; then
-        echo "Found a manual conda installation. Creating a project-specific virtual environment."
-        "${cmd_conda}" env create --force -f "${env_config_path}"
-    else
-        echo "Trying a system conda installation. Creating a project-specific virtual environment."
-        conda env create --force -f "${env_config_path}" 2>/dev/null
-    fi
-
-    exit 0
-else
-    echo "${0} does not support ${ID_LIKE} yet."
+if [ ! "${ID_LIKE}" = debian ] && [ ! "${ID}" = debian ]; then
+    print_warning "${0} does not support '${ID_LIKE}'/'${ID}' yet."
     exit 1
 fi
+
+## For system-wide or from-source installs that are not possible via apt
+sudo mkdir -p "${opt_dir}"
+sudo chown "${USER}:" "${opt_dir}"
+
+## Source the configurations for our dependencies
+. deps.sh
+
+echo "The user will now be prompted to install the following dependencies and optional features:
+  Binary packages (via apt-get), Clang, Boost, Docker, CUDA, OpenCV, eigen, Vulkan,
+  gtest, qemu, OpenXR-SDK, gtsam, opengv, DBoW2, Kimera-RPGO, Conda (miniconda3)
+" # End echo
+
+if y_or_n "Add apt-get sources list/keys and install necessary packages"; then
+    if y_or_n "^^^^  Also install Docker (docker-ce) for local CI/CD debugging support"; then
+        export use_docker="yes"
+    fi
+    pmt_msg_warn_cuda="Also automate install of CUDA 11 (cuda) for GPU plugin support on Ubuntu (_only_!)"
+    pmt_msg_warn_cuda+="\n(This script will _not_ install the package on non-Ubuntu distributions, "
+    pmt_msg_warn_cuda+="or if a supported GPU is not found)"
+    if y_or_n "^^^^  ${pmt_msg_warn_cuda}"; then
+        export use_cuda="yes"
+    fi
+
+    if [ "${enable_quiet}" = "no" ]; then
+        . "${script_path_apt}"
+    else
+        . "${script_path_apt}" >/dev/null
+    fi
+fi
+
+prompt_install \
+    "${dep_name_clang}" \
+    "${deps_log_dir}" \
+    "${script_path_clang}" \
+    "${parent_dir_clang}" \
+    "${dep_prompt_clang}" \
+    "${dep_ver_clang}"
+
+prompt_install \
+    "${dep_name_boost}" \
+    "${deps_log_dir}" \
+    "${script_path_boost}" \
+    "${parent_dir_boost}" \
+    "${dep_prompt_boost}" \
+    "${dep_ver_boost}"
+
+prompt_install \
+    "${dep_name_opencv}" \
+    "${deps_log_dir}" \
+    "${script_path_opencv}" \
+    "${parent_dir_opencv}" \
+    "${dep_prompt_opencv}" \
+    "${dep_ver_opencv}"
+
+prompt_install \
+    "${dep_name_eigen}" \
+    "${deps_log_dir}" \
+    "${script_path_eigen}" \
+    "${parent_dir_eigen}" \
+    "${dep_prompt_eigen}" \
+    "${dep_ver_eigen}"
+
+prompt_install \
+    "${dep_name_vulkan}" \
+    "${deps_log_dir}" \
+    "${script_path_vulkan}" \
+    "${parent_dir_vulkan}" \
+    "${dep_prompt_vulkan}" \
+    "${dep_ver_vulkan}"
+
+prompt_install \
+    "${dep_name_gtest}" \
+    "${deps_log_dir}" \
+    "${script_path_gtest}" \
+    "${parent_dir_gtest}" \
+    "${dep_prompt_gtest}" \
+    "${dep_ver_gtest}"
+
+prompt_install \
+    "${dep_name_qemu}" \
+    "${deps_log_dir}" \
+    "${script_path_qemu}" \
+    "${parent_dir_qemu}" \
+    "${dep_prompt_qemu}" \
+    "${dep_ver_qemu}"
+
+# if [ ! -d Vulkan-Loader ]; then
+#   echo "Next: Install Vulkan Loader from source"
+#   if y_or_n; then
+#       git clone https://github.com/KhronosGroup/Vulkan-Loader.git
+#       mkdir -p Vulkan-Headers/build && cd Vulkan-Headers/build
+#       ../scripts/update_deps.py
+#       cmake -C helper.cmake ..
+#       cmake --build .
+#       cd ../..
+#   fi
+# fi
+
+prompt_install \
+    "${dep_name_openxr}" \
+    "${deps_log_dir}" \
+    "${script_path_openxr}" \
+    "${parent_dir_openxr}" \
+    "${dep_prompt_openxr}" \
+    "${dep_ver_openxr}"
+
+prompt_install \
+    "${dep_name_gtsam}" \
+    "${deps_log_dir}" \
+    "${script_path_gtsam}" \
+    "${parent_dir_gtsam}" \
+    "${dep_prompt_gtsam}" \
+    "${dep_ver_gtsam}"
+
+prompt_install \
+    "${dep_name_opengv}" \
+    "${deps_log_dir}" \
+    "${script_path_opengv}" \
+    "${parent_dir_opengv}" \
+    "${dep_prompt_opengv}" \
+    "${dep_ver_opengv}"
+
+prompt_install \
+    "${dep_name_dbow2}" \
+    "${deps_log_dir}" \
+    "${script_path_dbow2}" \
+    "${parent_dir_dbow2}" \
+    "${dep_prompt_dbow2}" \
+    "${dep_ver_dbow2}"
+
+prompt_install \
+    "${dep_name_kimera_rpgo}" \
+    "${deps_log_dir}" \
+    "${script_path_kimera_rpgo}" \
+    "${parent_dir_kimera_rpgo}" \
+    "${dep_prompt_kimera_rpgo}" \
+    "${dep_ver_kimera_rpgo}"
+
+prompt_install \
+    "${dep_name_conda}" \
+    "${deps_log_dir}" \
+    "${script_path_conda}" \
+    "${parent_dir_conda}" \
+    "${dep_prompt_conda}" \
+    "${dep_ver_conda}"
+
+## Load new library paths
+sudo ldconfig
+
+### Virtual environment creation ###
+
+echo "Attempting to create a virtual environment configuration via conda ..."
+
+## Check for a previous conda installation
+detect_dependency "${dep_name_conda}" "${deps_log_dir}"
+if [ "$?" -eq 0 ]; then
+    dep_log_path_conda="${dep_log_path}"
+    src_dir_conda="${src_dir}"
+    echo "Found conda installation log '${dep_log_path_conda}' : dir <- '${src_dir_conda}'"
+else
+    dep_log_path_conda="${dep_log_path}"
+    src_dir_conda="${parent_dir_conda}/${dep_name_conda}"
+
+    dep_missing_msg_conda="Installation log for conda not found at '${dep_log_path_conda}'."
+    dep_missing_msg_conda+="\n  Conda may have been installed without this script (or an older version)."
+
+    print_warning "${dep_missing_msg_conda}"
+
+    if ! y_or_n "Try to create a Python environment configuration anyway?"; then
+        echo "This was the last step. Exiting early."
+        exit 0
+    fi
+
+    ## Conda may have been installed without this script
+    echo "Assuming : dir <- '${src_dir_conda}'"
+fi
+
+env_config_parent_dir = $(dirname "${env_config_path}")
+if [ ! -d "${env_config_parent_dir}" ]; then
+    mkdir -p "${env_config_parent_dir}"
+fi
+
+cmd_conda="${src_dir_conda}/bin/conda"
+if [ -f "${cmd_conda}" ]; then
+    echo "Found a manual conda installation. Creating a project-specific virtual environment."
+    "${cmd_conda}" env create --force -f "${env_config_path}"
+else
+    echo "Trying a system conda installation. Creating a project-specific virtual environment."
+    conda env create --force -f "${env_config_path}" 2>/dev/null
+fi
+
+exit 0
